Split tests into multiple files for maintainability and clarity
