+++ sd_food_bank_ai_bot/admin_panel/tests/__init__.py
@@ -1,0 +1,2 @@
+from .admin_panel_faq_tests import *
+from .phone_service_faq_tests import *
